Update makefile
As is, one needs to create ./bin and ./objs directories at the "root" (top-most) directory in oder for make to work correctly. The proposed changes added two lines that will create those directories for the user. Note that the -p switch is there to create a the path if it does not exist. In the man mkdir pages, one can see the description of -p is as follow: no error if existing, make parent directories as needed.